requirements.txt
python-dotenv
groq
mistralai
streamlit
matplotlib
plotly
pandas
pillow